Adjust code placement and add tests
